test_unit: do_test_01 succeeds

+++ test_unit.py
@@ -5,45 +5,87 @@ import numpy as np
 import subprocess
 import psycopg2
 
-from grid import db_def
+from grid import db_def, db_lib, insert_lib
 
 #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 # Error Handling and Logging
 
-# set up logging to file - see previous section for more details
 logging.basicConfig(level=logging.INFO,
                     format='%(asctime)s %(name)-12s %(levelname)-8s %(message)s',
                     datefmt='%m-%d %H:%M',
                     # filename='test_unit.log',
                     # filemode='w'
                     )
-# define a Handler which writes INFO messages or higher to the sys.stderr
-# console = logging.StreamHandler()
-# console.setLevel(logging.INFO)
-# set a format which is simpler for console use
 formatter = logging.Formatter('%(levelname)-8s: %(name)-12s: %(message)s')
-# tell the handler to use this format
-# console.setFormatter(formatter)
-# # add the handler to the root logger
-# logging.getLogger('').addHandler(console)
-
 logger = logging.getLogger(__name__)
 #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-def do_test_01():
 
-  logger.info('do_test_01')
+#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+def do_test_01(dbname):
 
-  return None
+  logger.info('do_test_01: Insert & retrieve a row into tracks, and ensure single-precision identicality.')
+
+  orig_id    = 1001
+  orig_M_ini = 12.3456
+  orig_fov   = 0.123
+  orig_Z     = 0.045
+  orig_logD  = 98.7654
+  tup        = (orig_id, orig_M_ini, orig_fov, orig_Z, orig_logD)
+
+  cmnd       = 'insert into tracks (id, M_ini, fov, Z, logD) values (%s,%s,%s,%s,%s)'
+  with db_def.grid_db(dbname=dbname) as the_db:
+    the_db.execute_one(cmnd, tup)
+
+  cmnd       = 'select id, M_ini, fov, Z, logD from tracks'
+  with db_def.grid_db(dbname=dbname) as the_db:
+    the_db.execute_one(cmnd, None)
+    result   = the_db.fetch_one()
+  res_id     = result[0]
+  res_M_ini  = result[1]
+  res_fov    = result[2]
+  res_Z      = result[3]
+  res_logD   = result[4]
+
+  # Define a local function to test each column value, and log the result properly
+  def make_local_test(attribute, original, retrieved):  
+    try:
+      assert original == retrieved
+      logger.info('    ... "{0}" OK'.format(attribute))
+      return True
+    except AssertionError:
+      logger.error('   ... "{0}" failed'.format(attribute))
+      return False
+
+  test_id    = make_local_test('id', orig_id, res_id)
+  test_M_ini = make_local_test('M_ini', orig_M_ini, res_M_ini)
+  test_fov   = make_local_test('fov', orig_fov, res_fov)
+  test_Z     = make_local_test('Z', orig_Z, res_Z)
+  test_logD  = make_local_test('logD', orig_logD, res_logD)
+
+  test       = all([test_id, test_M_ini, test_fov, test_Z, test_logD])
+  if test:
+    logger.info('do_test_01: All five columns retrieved successfully')
+  else:
+    logger.error('do_test_01: At least one column was not retrieved successfully')
+
+  cmnd       = 'delete from tracks'
+  with db_def.grid_db(dbname=dbname) as the_db:
+    the_db.execute_one(cmnd, None)
+
+  return test
 
 #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 def make_table_tracks(dbname):
   """
   Create the "tracks" table identical to the "grid.tracks" table
   """
-  tbl =  'create table grid.tracks ( \
+  if not db_def.exists(dbname=dbname): 
+    logger.error('make_table_tracks: Database "{0}" does not exist'.format(dbname))
+    sys.exit(1)
+
+  tbl =  'create table tracks ( \
           id             serial, \
           M_ini          real not null, \
           fov            real not null, \
@@ -56,24 +98,43 @@ def make_table_tracks(dbname):
           constraint positive_Z check (Z > 0), \
           constraint positive_log_D check (logD >= 0) \
         ); \
-        create index index_track_id on grid.tracks (id asc);  \
-        create index index_M_ini on grid.tracks (M_ini asc); \
-        create index index_fov on grid.tracks (fov asc); \
-        create index index_Z on grid.tracks (Z asc); \
-        create index index_log_D on grid.tracks (logD asc);'
+        create index index_track_id on tracks (id asc);  \
+        create index index_M_ini on tracks (M_ini asc); \
+        create index index_fov on tracks (fov asc); \
+        create index index_Z on tracks (Z asc); \
+        create index index_log_D on tracks (logD asc);'.format(dbname)
 
-  my_db = db_def.grid_db(dbname=dbname)
-  with temp_db as my_db:
-    temp_db.execute_one(tbl)
+  with db_def.grid_db(dbname=dbname) as my_db:
+    my_db.execute_one(tbl, None)
 
   logger.info('make_table_tracks: the "tracks" table created in database "{0}".'.format(dbname))
 
+#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+def make_schema(dbname):
+  """
+  Create a schema for the database, like the following: "create schema grid;"
+  """
+  if db_def.exists(dbname): return True
+
+  schema = 'create schema {0}'.format(dbname)
+  with db_def.grid_db(dbname=dbname) as the_db:
+    the_db.execute_one(schema, None)
+  logger.info('make_schema: "{0}" done'.format(schema))
+
+  return True
 
 #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 def create_test_database(dbname):
   """
   Create a new database by calling shell commands.
+  Returns false if failed to create the database, and returns True if the database already existed 
+  or successfully created
   """
+  present = db_def.exists(dbname)
+  if present:
+    logger.warning('create_test_database: Database "{0}" already existed.'.format(dbname))
+    return True
+
   cmnd    = 'createdb {0}'.format(dbname)
   execute = subprocess.Popen(cmnd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
   stdout  = execute.stdout.read().rstrip('\n')
@@ -83,14 +144,23 @@ def create_test_database(dbname):
     logger.error('create_test_database: stdout: {0}'.format(stdout))
     logger.error('create_test_database: stderr: {0}'.format(stderr))
     logger.error('create_test_database: failed to create the database')
+    return False
   else:
     logger.info('create_test_database: database "{0}" created'.format(dbname))
+    return True
 
 #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 def drop_test_database(dbname):
   """
   Delete/drop the test database by calling a shell command
+  Returns False if fails to drop the dataase, and returns True if the database was already not there,
+  or when the database is successfully dropped.
   """
+  present = db_def.exists(dbname)
+  if not present:
+    logger.warning('drop_test_database: Database "{0}" does not exist anyway. Cannot drop it.')
+    return True
+
   cmnd    = 'dropdb {0}'.format(dbname)
   execute = subprocess.Popen(cmnd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
   stdout  = execute.stdout.read().rstrip('\n')
@@ -100,8 +170,10 @@ def drop_test_database(dbname):
     logger.error('drop_test_database: stdout: {0}'.format(stdout))
     logger.error('drop_test_database: stderr: {0}'.format(stderr))
     logger.error('drop_test_database: failed to drop the database "{0}"'.format(dbname))
+    return False
   else:
     logger.info('drop_test_database: successfully droped the database "{0}"'.format(dbname))
+    return True
 
 #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 def main():
@@ -112,11 +184,21 @@ def main():
 
   my_db   = 'test_grid' 
 
-  create_test_database(dbname=my_db)
+  status  = create_test_database(dbname=my_db)
+  if status is not True:
+    logger.error('main: create_test_database failed')
+    return status
+
+  status  = make_schema(dbname=my_db)
+
   make_table_tracks(dbname=my_db)
 
-  # do_test_01()
-  drop_test_database(dbname=my_db)
+  status  = do_test_01(dbname=my_db)
+
+  status  = drop_test_database(dbname=my_db)
+  if status is not True:
+    logger.error('main: drop_test_db failed')
+    return status
 
 #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 if __name__ == '__main__':
